+++ events/src/main/java/mk/polarcape/repository/EmployeeRepository.java
@@ -10,5 +10,6 @@ public interface EmployeeRepository {
 	Employee save(Employee task);
 	int delete(Long id);
 	List<Employee> getInvitedEmployeeForEventId(Long id);
-
+	Employee login(String username, String pass);
+	
 }
