--- events/src/main/java/mk/polarcape/repository/LoginRepository.java
@@ -1,8 +1,0 @@
-package mk.polarcape.repository;
-
-import mk.polarcape.model.Login;
-
-public interface LoginRepository {
-
-	Login login(String username, String pass);
-}
