--- events/src/main/java/mk/polarcape/service/LoginService.java
@@ -1,7 +1,0 @@
-package mk.polarcape.service;
-
-import mk.polarcape.model.Login;
-
-public interface LoginService {
-	Login login(String username, String pass);
-}
